Refactor commands into separate files
