+++ Managers/WebSocketServerManager.cs
@@ -13,6 +13,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Axys.Utilities;
 using Rhino.Commands;
+using Axys.Commands;
 
 namespace Axys
 {
