+++ Models/RhinoObjectInfo.cs
@@ -1,0 +1,10 @@
+using System;
+namespace Axys.Models
+{
+    public class RhinoObjectInfo
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; }
+        public string ObjectType { get; set; }
+    }
+}
